LuceneIndex

+++ src/main/java/com/xlongwei/search/LucenePlus.java
@@ -1,25 +1,12 @@
 package com.xlongwei.search;
 
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
-import java.io.InputStreamReader;
-import java.io.OutputStreamWriter;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.networknt.config.Config;
 import com.networknt.server.ShutdownHookProvider;
-import com.networknt.utility.StringUtils;
 
 import org.apache.lucene.analysis.Analyzer;
 import org.apache.lucene.analysis.standard.StandardAnalyzer;
@@ -29,8 +16,12 @@ import org.apache.lucene.index.DirectoryReader;
 import org.apache.lucene.index.IndexWriter;
 import org.apache.lucene.index.IndexWriterConfig;
 import org.apache.lucene.search.IndexSearcher;
-import org.apache.lucene.store.FSDirectory;
+import org.apache.lucene.store.Directory;
+import org.apache.lucene.store.FSLockFactory;
+import org.apache.lucene.store.LockFactory;
+import org.apache.lucene.store.MMapDirectory;
 import org.apache.lucene.store.NIOFSDirectory;
+import org.apache.lucene.store.NoLockFactory;
 
 import io.undertow.util.FileUtils;
 import lombok.extern.slf4j.Slf4j;
@@ -42,28 +33,8 @@ import lombok.extern.slf4j.Slf4j;
 @SuppressWarnings({ "rawtypes" })
 public class LucenePlus implements ShutdownHookProvider {
 
-    private static Map<String, Object> config = Config.getInstance().getJsonMapConfig("lucene");
-    private static String index = (String) config.get("index");
-    private static String indices = "indices.json";
     private static Map<String, IndexWriter> writers = new HashMap<>();
     private static Map<String, IndexSearcher> searchers = new HashMap<>();
-    private static Map<String, List<LuceneField>> fields = new HashMap<>();
-
-    static {
-        Path path = Paths.get(index, indices);
-        File file = path.toFile();
-        try (BufferedReader reader = file.exists()
-                ? new BufferedReader(new InputStreamReader(new FileInputStream(file), StandardCharsets.UTF_8))
-                : null) {
-            String json = file.exists() ? reader.readLine() : StringUtils.EMPTY;
-            json = StringUtils.isNotBlank(json) ? json : "{}";
-            fields = Config.getInstance().getMapper().readValue(json,
-                    new TypeReference<Map<String, List<LuceneField>>>() {
-                    });
-        } catch (Exception e) {
-            log.warn("fail to load indices", e);
-        }
-    }
 
     public static IndexWriter getWriter(String name) throws Exception {
         IndexWriter writer = writers.get(name);
@@ -71,7 +42,7 @@ public class LucenePlus implements ShutdownHookProvider {
             synchronized (writers) {
                 writer = writers.get(name);
                 if (writer == null) {
-                    FSDirectory directory = NIOFSDirectory.open(Paths.get(index, name));
+                    Directory directory = getDirectory(name);
                     writer = new IndexWriter(directory, new IndexWriterConfig(getAnalyzer(name)));
                     writers.put(name, writer);
                 }
@@ -86,8 +57,10 @@ public class LucenePlus implements ShutdownHookProvider {
             synchronized (searchers) {
                 searcher = searchers.get(name);
                 if (searcher == null) {
-                    FSDirectory directory = NIOFSDirectory.open(Paths.get(index, name));
-                    DirectoryReader reader = DirectoryReader.open(directory);
+                    LuceneIndex index = LuceneIndex.index(name);
+                    boolean realtime = index != null ? index.isRealtime() : false;
+                    DirectoryReader reader = realtime ? DirectoryReader.open(getWriter(name))
+                            : DirectoryReader.open(getDirectory(name));
                     searcher = new IndexSearcher(reader);
                     searchers.put(name, searcher);
                 }
@@ -100,9 +73,33 @@ public class LucenePlus implements ShutdownHookProvider {
         return new StandardAnalyzer();
     }
 
+    public static Directory getDirectory(String name) throws Exception {
+        Path path = LuceneIndex.path(name);
+        LuceneIndex index = LuceneIndex.index(name);
+        LockFactory lockFactory = NoLockFactory.INSTANCE;
+        if (index != null && index.getDirectory() != null) {
+            if ("fs".equals(index.getLockFactory())) {
+                lockFactory = FSLockFactory.getDefault();
+            }
+            switch (index.getDirectory()) {
+                case "nio":
+                    return NIOFSDirectory.open(path, lockFactory);
+                case "mmap":
+                    return MMapDirectory.open(path, lockFactory);
+            }
+        }
+        String systemName = System.getProperty("os.name").toLowerCase();
+        return systemName.contains("win") ? MMapDirectory.open(path, lockFactory)
+                : NIOFSDirectory.open(path, lockFactory);
+    }
+
     public static boolean close(String name) throws Exception {
         IndexWriter writer = writers.get(name);
         if (writer != null) {
+            // 近实时搜索不必关闭
+            if (LuceneIndex.realtime(name)) {
+                return true;
+            }
             writer.close();
             writers.remove(name);
             // 关闭索引之后重新获取searcher
@@ -111,20 +108,33 @@ public class LucenePlus implements ShutdownHookProvider {
         return writer != null;
     }
 
+    // 关闭searcher，不关闭近实时索引
     public static boolean closeSearcher(String name) throws Exception {
         IndexSearcher searcher = searchers.remove(name);
         if (searcher != null) {
-            searcher.getIndexReader().close();
+            if (!LuceneIndex.realtime(name)) {
+                searcher.getIndexReader().close();
+            }
             return true;
         }
         return false;
     }
 
     public static boolean drop(String name) throws Exception {
-        close(name);
-        Path path = Paths.get(index, name);
-        FileUtils.deleteRecursive(path);
-        return fields(name, null);
+        boolean index = LuceneIndex.index(name, null);
+        if (index) {
+            IndexSearcher searcher = searchers.remove(name);
+            if (searcher != null && !LuceneIndex.realtime(name)) {
+                searcher.getIndexReader().close();
+            }
+            IndexWriter writer = writers.remove(name);
+            if (writer != null) {
+                writer.close();
+            }
+            Path path = LuceneIndex.path(name);
+            FileUtils.deleteRecursive(path);
+        }
+        return index;
     }
 
     /** {add:[{name:value}]} */
@@ -132,7 +142,7 @@ public class LucenePlus implements ShutdownHookProvider {
         if (map == null || map.isEmpty()) {
             return false;
         }
-        List<LuceneField> fields = fields(name);
+        List<LuceneField> fields = LuceneIndex.fields(name);
         if (fields == null || fields.isEmpty()) {
             return false;
         }
@@ -162,44 +172,6 @@ public class LucenePlus implements ShutdownHookProvider {
             return true;
         }
         return false;
-    }
-
-    /**
-     * @param fields null 删除索引，non-null 初始化索引，多次提交时不处理
-     */
-    public static boolean fields(String name, List<LuceneField> fields) throws Exception {
-        Path path = Paths.get(index, indices);
-        if (fields == null) {
-            List<LuceneField> remove = LucenePlus.fields.remove(name);
-            if (remove != null) {
-                BufferedWriter writer = new BufferedWriter(
-                        new OutputStreamWriter(new FileOutputStream(path.toFile()), StandardCharsets.UTF_8));
-                writer.write(Config.getInstance().getMapper().writeValueAsString(LucenePlus.fields));
-                writer.close();
-                return true;
-            }
-        } else {
-            if (!LucenePlus.fields.containsKey(name)) {
-                LucenePlus.fields.put(name, fields);
-                BufferedWriter writer = new BufferedWriter(
-                        new OutputStreamWriter(new FileOutputStream(path.toFile()), StandardCharsets.UTF_8));
-                writer.write(Config.getInstance().getMapper().writeValueAsString(LucenePlus.fields));
-                writer.close();
-                return true;
-            }
-        }
-        return false;
-    }
-
-    /**
-     * 获取索引的字段列表，需要先打开索引
-     * 
-     * @param name
-     * @return
-     */
-    public static List<LuceneField> fields(String name) {
-        List<LuceneField> list = fields.get(name);
-        return list != null ? list : Collections.emptyList();
     }
 
     @Override
